refactor(tests): organize backend suites by domain
