+++ backend/internal/api/connector_actions_test.go
@@ -2,11 +2,6 @@ package api
 
 import (
 	"context"
-	"database/sql"
-	"fmt"
-	"net/http"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -17,10 +12,7 @@ import (
 	sshconnector "github.com/aipermission/aipermission/backend/internal/connectors/ssh"
 	"github.com/aipermission/aipermission/backend/internal/connectors/ssh/sshkeys"
 	"github.com/aipermission/aipermission/backend/internal/connectortargets"
-	dbpkg "github.com/aipermission/aipermission/backend/internal/db"
-	historypkg "github.com/aipermission/aipermission/backend/internal/history"
 	"github.com/aipermission/aipermission/backend/internal/tokens"
-	"github.com/aipermission/aipermission/backend/internal/vault"
 )
 
 func TestRuntimePrepareConnectorActionUsesSSHConnectorProfile(t *testing.T) {
@@ -215,1068 +207,4 @@ func TestConnectorApprovalContextHashesConnectorAndActionDefinition(t *testing.T
 	if renamedTokenHash != baseHash {
 		t.Fatalf("token label changes should not change approval hash")
 	}
-}
-
-func TestCallConnectorActionBlocksMissingPermission(t *testing.T) {
-	database := openAPITestDB(t)
-	secretVault := openAPITestVault(t)
-	runtime := connectorActionTestRuntime(t, database, secretVault)
-	server := &Server{}
-	store := connectortargets.NewStore(database)
-	tokenID := insertAPITestToken(t, database)
-	target, profile := createAPITestPostgresTargetProfile(t, store, secretVault)
-
-	result, err := server.callConnectorAction(context.Background(), runtime, connectorActionCall{
-		Source:     commandRequestSourceMCP,
-		TokenID:    tokenID,
-		TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-		ActionName: postgresconnector.ActionQueryReadonly,
-		Input:      map[string]any{"sql": "select 1"},
-		Reason:     "smoke",
-	})
-	if err != nil {
-		t.Fatalf("call connector action: %v", err)
-	}
-	if result.Result.Status != connectors.ResultBlocked || result.Request.Status != connectors.ResultBlocked {
-		t.Fatalf("expected blocked result/request, got %#v", result)
-	}
-	if result.Request.CompletedAt == nil || result.Request.Error == "" {
-		t.Fatalf("blocked request should be terminal with error: %#v", result.Request)
-	}
-}
-
-func TestCallConnectorActionCreatesPendingApproval(t *testing.T) {
-	database := openAPITestDB(t)
-	secretVault := openAPITestVault(t)
-	runtime := connectorActionTestRuntime(t, database, secretVault)
-	server := &Server{}
-	store := connectortargets.NewStore(database)
-	tokenID := insertAPITestToken(t, database)
-	target, profile := createAPITestPostgresTargetProfile(t, store, secretVault)
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       tokenID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-	}); err != nil {
-		t.Fatalf("set action permission: %v", err)
-	}
-
-	result, err := server.callConnectorAction(context.Background(), runtime, connectorActionCall{
-		Source:     commandRequestSourceMCP,
-		TokenID:    tokenID,
-		TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-		ActionName: postgresconnector.ActionQueryReadonly,
-		Input:      map[string]any{"sql": "select 1", "max_rows": 5},
-		Reason:     "inspect one row",
-	})
-	if err != nil {
-		t.Fatalf("call connector action: %v", err)
-	}
-	if result.Result.Status != connectors.ResultApprovalPending || result.Request.Status != connectors.ResultApprovalPending {
-		t.Fatalf("expected pending result/request, got %#v", result)
-	}
-	if result.Request.EncryptedPayloadJSON == "" || result.Request.ApprovalContextHash == "" {
-		t.Fatalf("pending request missing encrypted payload/context: %#v", result.Request)
-	}
-	if result.Result.Handles.RequestID != result.Request.ID || result.Result.Handles.FollowupTool == "" {
-		t.Fatalf("pending result missing followup handle: %#v", result.Result)
-	}
-}
-
-func TestRunLocalConnectorActionCreatesManualHistory(t *testing.T) {
-	database := openAPITestDB(t)
-	secretVault := openAPITestVault(t)
-	registry := connectors.NewRegistry()
-	if err := registry.Register(localActionTestConnector{}); err != nil {
-		t.Fatalf("register local test connector: %v", err)
-	}
-	runtime := &databaseRuntime{
-		database: database,
-		vault:    secretVault,
-		tokens:   tokens.NewStore(database),
-		registry: registry,
-	}
-	server := &Server{}
-	store := connectortargets.NewStore(database)
-	target, err := store.CreateTarget(context.Background(), connectortargets.CreateTargetInput{
-		ConnectorKind: localActionTestConnectorKind,
-		Name:          "local-target",
-		Config:        map[string]any{},
-	})
-	if err != nil {
-		t.Fatalf("create local target: %v", err)
-	}
-	profile, err := store.CreateCredentialProfile(context.Background(), connectortargets.CreateCredentialProfileInput{
-		TargetID:      target.ID,
-		ConnectorKind: localActionTestConnectorKind,
-		Kind:          "default",
-		Label:         "main",
-		Public:        map[string]any{},
-	})
-	if err != nil {
-		t.Fatalf("create local profile: %v", err)
-	}
-
-	result, err := server.runLocalConnectorAction(context.Background(), runtime, connectorActionCall{
-		TargetRef:  connectortargets.ConnectorTargetRef(localActionTestConnectorKind, target.ID, profile.ID),
-		ActionName: "echo",
-		Input:      map[string]any{"value": "hello"},
-		Reason:     "manual console smoke",
-	})
-	if err != nil {
-		t.Fatalf("run local connector action: %v", err)
-	}
-	if result.Request.TokenID != nil || result.Request.Source != commandRequestSourceManual || result.Request.Status != connectors.ResultCompleted {
-		t.Fatalf("unexpected local request: %#v", result.Request)
-	}
-	if result.Result.Output.(map[string]any)["echo"] != "hello" {
-		t.Fatalf("unexpected output: %#v", result.Result.Output)
-	}
-	var historyCount int
-	if err := database.QueryRow(`
-		SELECT COUNT(*)
-		FROM history_entries
-		WHERE source_ref_type = ? AND source_ref_id = ? AND source = 'manual' AND connector_kind = ?`,
-		historypkg.SourceConnectorActionRequest,
-		result.Request.ID,
-		localActionTestConnectorKind,
-	).Scan(&historyCount); err != nil {
-		t.Fatalf("read local action history: %v", err)
-	}
-	if historyCount != 1 {
-		t.Fatalf("expected one local action history row, got %d", historyCount)
-	}
-
-	completedWithHandle, err := server.runLocalConnectorAction(context.Background(), runtime, connectorActionCall{
-		TargetRef:  connectortargets.ConnectorTargetRef(localActionTestConnectorKind, target.ID, profile.ID),
-		ActionName: "echo",
-		Input:      map[string]any{"value": "with-handle"},
-		Reason:     "capture completed session handle",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if completedWithHandle.Request.Status != connectors.ResultCompleted ||
-		completedWithHandle.Request.SessionID == nil || *completedWithHandle.Request.SessionID != 123 ||
-		completedWithHandle.Request.SessionGeneration == nil || *completedWithHandle.Request.SessionGeneration != 456 {
-		t.Fatalf("completed action handle was not persisted: %#v", completedWithHandle.Request)
-	}
-
-	incompleteHandle, err := server.runLocalConnectorAction(context.Background(), runtime, connectorActionCall{
-		TargetRef:  connectortargets.ConnectorTargetRef(localActionTestConnectorKind, target.ID, profile.ID),
-		ActionName: "echo",
-		Input:      map[string]any{"value": "incomplete-handle"},
-		Reason:     "reject incomplete session handle",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if incompleteHandle.Request.Status != connectors.ResultFailed ||
-		incompleteHandle.Result.Status != connectors.ResultFailed ||
-		!strings.Contains(incompleteHandle.Result.Error, "incomplete session handle") {
-		t.Fatalf("incomplete session handle should fail terminally: %#v", incompleteHandle)
-	}
-
-	classifiedFailure, err := server.runLocalConnectorAction(context.Background(), runtime, connectorActionCall{
-		TargetRef:  connectortargets.ConnectorTargetRef(localActionTestConnectorKind, target.ID, profile.ID),
-		ActionName: "echo",
-		Input:      map[string]any{"value": "classified-error"},
-		Reason:     "preserve a stable connector error code",
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	output, _ := classifiedFailure.Request.Output.(map[string]any)
-	if classifiedFailure.Request.Status != connectors.ResultFailed || output["code"] != "fixture_failure" || classifiedFailure.Result.Output.(map[string]any)["code"] != "fixture_failure" {
-		t.Fatalf("classified connector error code was not persisted: %#v", classifiedFailure)
-	}
-}
-
-func TestInsertConnectorActionRequestRedactsDisplayedInputOnly(t *testing.T) {
-	database := openAPITestDB(t)
-	secretVault := openAPITestVault(t)
-	runtime := connectorActionTestRuntime(t, database, secretVault)
-	server := &Server{}
-	if _, err := insertRedactionRule(t.Context(), runtime, redactionRuleRequest{
-		Name: "approval preview token", Pattern: `internal_[a-z0-9]+`, Enabled: true,
-	}); err != nil {
-		t.Fatalf("insert custom redaction rule: %v", err)
-	}
-	store := connectortargets.NewStore(database)
-	tokenID := insertAPITestToken(t, database)
-	target, profile := createAPITestPostgresTargetProfile(t, store, secretVault)
-	targetView, profileView, err := store.ResolveConnectorActionTarget(context.Background(), connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID))
-	if err != nil {
-		t.Fatalf("resolve target/profile: %v", err)
-	}
-	rawInput := map[string]any{
-		"sql":            "select 'password=super-secret' as value",
-		"access_token":   "raw-access-token",
-		"opaque_message": "arbitrary-publish-content",
-		"nested":         map[string]any{"authorization": "Bearer raw-bearer-token"},
-	}
-	prepared := actions.PreparedRequest{
-		Target:  targetView,
-		Profile: profileView,
-		ActionDefinition: connectors.ActionDefinition{
-			Name:                 postgresconnector.ActionQueryReadonly,
-			SensitiveInputFields: []string{"access_token", "opaque_message", "client-secret"},
-		},
-		Action: connectors.PreparedAction{
-			ConnectorKind: postgresconnector.Kind,
-			TargetRef:     targetView.Ref,
-			ProfileID:     profile.ID,
-			ActionName:    postgresconnector.ActionQueryReadonly,
-			Preview:       map[string]any{"body": "password=visible-for-approval internal_abc123", "opaque_message": "exact-sensitive-preview", "client_secret": "hyphen-normalized-preview-secret"},
-			Payload:       rawInput,
-		},
-		Requested: actions.PrepareRequest{
-			Source:     commandRequestSourceMCP,
-			TargetRef:  targetView.Ref,
-			ActionName: postgresconnector.ActionQueryReadonly,
-			Input:      rawInput,
-			Reason:     "Bearer raw-reason-token password=reason-secret",
-		},
-	}
-
-	request, err := server.insertConnectorActionRequest(context.Background(), runtime, tokenID, prepared, connectortargets.ActionPermission{}, connectors.ResultApprovalPending, "")
-	if err != nil {
-		t.Fatalf("insert connector action request: %v", err)
-	}
-	var inputJSON string
-	var previewJSON string
-	var reason string
-	var encryptedPayload string
-	if err := database.QueryRow(`
-			SELECT input_json, preview_json, reason, encrypted_payload_json
-		FROM connector_action_requests
-		WHERE id = ?`,
-		request.ID,
-	).Scan(&inputJSON, &previewJSON, &reason, &encryptedPayload); err != nil {
-		t.Fatalf("read connector action request: %v", err)
-	}
-	for _, secret := range []string{"super-secret", "raw-access-token", "arbitrary-publish-content", "raw-bearer-token", "raw-reason-token", "reason-secret"} {
-		if strings.Contains(inputJSON, secret) || strings.Contains(reason, secret) {
-			t.Fatalf("persisted connector request leaked %q: input=%s reason=%s", secret, inputJSON, reason)
-		}
-	}
-	if !strings.Contains(inputJSON, `"access_token":"[REDACTED]"`) || !strings.Contains(inputJSON, `"opaque_message":"[REDACTED]"`) || !strings.Contains(inputJSON, `"authorization":"[REDACTED]"`) || !strings.Contains(inputJSON, `password=[REDACTED]`) {
-		t.Fatalf("input was not redacted as expected: %s", inputJSON)
-	}
-	if strings.Contains(previewJSON, "visible-for-approval") || strings.Contains(previewJSON, "internal_abc123") {
-		t.Fatalf("persisted display preview leaked exact approval content: %s", previewJSON)
-	}
-	if strings.Contains(previewJSON, "exact-sensitive-preview") || !strings.Contains(previewJSON, `"opaque_message":"[REDACTED]"`) {
-		t.Fatalf("persisted display preview did not apply sensitive input fields: %s", previewJSON)
-	}
-	if strings.Contains(previewJSON, "hyphen-normalized-preview-secret") || !strings.Contains(previewJSON, `"client_secret":"[REDACTED]"`) {
-		t.Fatalf("persisted display preview did not normalize declared sensitive fields: %s", previewJSON)
-	}
-	var historyInputJSON string
-	if err := database.QueryRow(`
-		SELECT input_json
-		FROM history_entries
-		WHERE source_ref_type = 'connector_action_request' AND source_ref_id = ?`,
-		request.ID,
-	).Scan(&historyInputJSON); err != nil {
-		t.Fatalf("read connector history input: %v", err)
-	}
-	if historyInputJSON != inputJSON {
-		t.Fatalf("history input drifted from redacted request input: history=%s request=%s", historyInputJSON, inputJSON)
-	}
-	mcpResponse := connectorActionRequestToMCPResponse(request)
-	approvalResponse := connectorActionApprovalItemFromRequest(request)
-	if mcpResponse.Input["access_token"] != "[REDACTED]" || approvalResponse.Input["access_token"] != "[REDACTED]" {
-		t.Fatalf("response input was not redacted: mcp=%#v approval=%#v", mcpResponse.Input, approvalResponse.Input)
-	}
-	if mcpResponse.Input["opaque_message"] != "[REDACTED]" || approvalResponse.Input["opaque_message"] != "[REDACTED]" {
-		t.Fatalf("action-declared sensitive input was not redacted: mcp=%#v approval=%#v", mcpResponse.Input, approvalResponse.Input)
-	}
-	exactApproval, err := connectorActionApprovalItemForResponse(runtime, request)
-	if err != nil {
-		t.Fatalf("build exact approval response: %v", err)
-	}
-	if exactApproval.Preview["body"] != "password=visible-for-approval internal_abc123" {
-		t.Fatalf("pending approval preview must match the exact prepared action: %#v", exactApproval.Preview)
-	}
-	if exactApproval.Preview["opaque_message"] != "exact-sensitive-preview" {
-		t.Fatalf("exact approval preview must preserve sensitive content inside the encrypted envelope: %#v", exactApproval.Preview)
-	}
-	if exactApproval.Preview["client_secret"] != "hyphen-normalized-preview-secret" {
-		t.Fatalf("exact approval preview must preserve normalized sensitive content inside the encrypted envelope: %#v", exactApproval.Preview)
-	}
-	redactedApproval := connectorActionApprovalItemFromRequest(request)
-	if redactedApproval.Preview["body"] == "password=visible-for-approval internal_abc123" {
-		t.Fatalf("approval list projection exposed exact preview: %#v", redactedApproval.Preview)
-	}
-	var decryptedPayload connectorActionExecutionEnvelope
-	if err := secretVault.DecryptJSON(encryptedPayload, &decryptedPayload); err != nil {
-		t.Fatalf("decrypt execution payload: %v", err)
-	}
-	if decryptedPayload.Input["access_token"] != "raw-access-token" || !strings.Contains(decryptedPayload.Input["sql"].(string), "super-secret") {
-		t.Fatalf("encrypted execution payload should preserve raw input: %#v", decryptedPayload)
-	}
-	if decryptedPayload.Input["opaque_message"] != "arbitrary-publish-content" {
-		t.Fatalf("encrypted execution payload should preserve action-sensitive input: %#v", decryptedPayload)
-	}
-	if decryptedPayload.Payload["access_token"] != "raw-access-token" || !strings.Contains(decryptedPayload.Payload["sql"].(string), "super-secret") {
-		t.Fatalf("encrypted execution payload should preserve raw action payload: %#v", decryptedPayload)
-	}
-	if decryptedPayload.ApprovalPreview["body"] != "password=visible-for-approval internal_abc123" {
-		t.Fatalf("encrypted approval preview changed: %#v", decryptedPayload.ApprovalPreview)
-	}
-	if !strings.Contains(decryptedPayload.Reason, "raw-reason-token") || !strings.Contains(decryptedPayload.Reason, "reason-secret") {
-		t.Fatalf("encrypted execution payload should preserve raw reason: %#v", decryptedPayload)
-	}
-}
-
-func TestRunningConnectorActionResponseRedactsOutput(t *testing.T) {
-	database := openAPITestDB(t)
-	secretVault := openAPITestVault(t)
-	runtime := connectorActionTestRuntime(t, database, secretVault)
-	server := &Server{}
-	store := connectortargets.NewStore(database)
-	tokenID := insertAPITestToken(t, database)
-	target, profile := createAPITestPostgresTargetProfile(t, store, secretVault)
-	request, err := store.InsertActionRequest(context.Background(), connectortargets.InsertActionRequestInput{
-		TokenID:       &tokenID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ConnectorKind: postgresconnector.Kind,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		Input:         map[string]any{"sql": "select 1"},
-		Status:        connectors.ResultRunning,
-	})
-	if err != nil {
-		t.Fatalf("insert running request: %v", err)
-	}
-	redacted := server.redactConnectorActionResult(context.Background(), runtime, connectors.ActionResult{
-		Status:      connectors.ResultRunning,
-		Output:      map[string]any{"rows": []map[string]any{{"session_token": "raw-token", "name": "safe"}}},
-		DisplayText: "Bearer raw-bearer-token",
-		Error:       "password=super-secret",
-	}, connectors.OutputHint{SensitiveFields: []string{"session_token"}})
-	response := connectorActionToMCPResponse(request, redacted)
-	payload := fmt.Sprint(response.Output, response.DisplayText, response.Error)
-	for _, secret := range []string{"raw-token", "raw-bearer-token", "super-secret"} {
-		if strings.Contains(payload, secret) {
-			t.Fatalf("running response leaked %q: %#v", secret, response)
-		}
-	}
-}
-
-func TestFinishConnectorActionRequestRedactsErrorAndHistory(t *testing.T) {
-	database := openAPITestDB(t)
-	secretVault := openAPITestVault(t)
-	runtime := connectorActionTestRuntime(t, database, secretVault)
-	server := &Server{}
-	store := connectortargets.NewStore(database)
-	tokenID := insertAPITestToken(t, database)
-	target, profile := createAPITestPostgresTargetProfile(t, store, secretVault)
-
-	request, err := store.InsertActionRequest(context.Background(), connectortargets.InsertActionRequestInput{
-		TokenID:              &tokenID,
-		TargetID:             target.ID,
-		ProfileID:            profile.ID,
-		ConnectorKind:        postgresconnector.Kind,
-		ActionName:           postgresconnector.ActionQueryReadonly,
-		Input:                map[string]any{"sql": "select 1"},
-		EncryptedPayloadJSON: "encrypted",
-		Status:               connectors.ResultRunning,
-	})
-	if err != nil {
-		t.Fatalf("insert action request: %v", err)
-	}
-	finished, err := server.finishConnectorActionRequest(
-		context.Background(),
-		runtime,
-		request.ID,
-		connectors.ResultFailed,
-		map[string]any{
-			"rows": []map[string]any{
-				{
-					"customer_secret": "visible-only-if-buggy",
-					"token":           "token-value",
-					"access_token":    "access-token-value",
-					"password_hash":   "password-hash-value",
-					"api_token_hash":  "api-token-hash-value",
-					"secret_value":    "secret-value",
-					"name":            "safe",
-				},
-			},
-		},
-		"",
-		"connect failed password=super-secret Bearer abcdefghijklmnopqrstuvwxyz",
-		connectors.OutputHint{SensitiveFields: []string{"customer_secret"}},
-	)
-	if err != nil {
-		t.Fatalf("finish action request: %v", err)
-	}
-	if strings.Contains(finished.Error, "super-secret") || strings.Contains(finished.Error, "abcdefghijklmnopqrstuvwxyz") {
-		t.Fatalf("finished error leaked secret: %q", finished.Error)
-	}
-	if !strings.Contains(finished.Error, "password=[REDACTED]") || !strings.Contains(finished.Error, "Bearer [REDACTED]") {
-		t.Fatalf("finished error was not redacted as expected: %q", finished.Error)
-	}
-	var historyError string
-	if err := database.QueryRow(`
-		SELECT error
-		FROM history_entries
-		WHERE source_ref_type = 'connector_action_request' AND source_ref_id = ?`,
-		request.ID,
-	).Scan(&historyError); err != nil {
-		t.Fatalf("read history error: %v", err)
-	}
-	if historyError != finished.Error {
-		t.Fatalf("history error drifted from finished request: history=%q finished=%q", historyError, finished.Error)
-	}
-	response := connectorActionToMCPResponse(finished, connectors.ActionResult{Status: connectors.ResultFailed, Error: finished.Error})
-	if strings.Contains(response.Error, "super-secret") || strings.Contains(response.Error, "abcdefghijklmnopqrstuvwxyz") {
-		t.Fatalf("mcp response leaked secret: %q", response.Error)
-	}
-	if response.Error != finished.Error {
-		t.Fatalf("mcp response error drifted from finished request: response=%q finished=%q", response.Error, finished.Error)
-	}
-	var outputJSON string
-	if err := database.QueryRow(`
-		SELECT output_json
-		FROM connector_action_requests
-		WHERE id = ?`,
-		request.ID,
-	).Scan(&outputJSON); err != nil {
-		t.Fatalf("read connector action output: %v", err)
-	}
-	for _, secret := range []string{"visible-only-if-buggy", "token-value", "access-token-value", "password-hash-value", "api-token-hash-value", "secret-value"} {
-		if strings.Contains(outputJSON, secret) {
-			t.Fatalf("structured connector output leaked sensitive field value %q: %s", secret, outputJSON)
-		}
-	}
-	for _, marker := range []string{`"customer_secret":"[REDACTED]"`, `"token":"[REDACTED]"`, `"access_token":"[REDACTED]"`, `"password_hash":"[REDACTED]"`, `"api_token_hash":"[REDACTED]"`, `"secret_value":"[REDACTED]"`} {
-		if !strings.Contains(outputJSON, marker) {
-			t.Fatalf("structured connector output missing redacted marker %s: %s", marker, outputJSON)
-		}
-	}
-	if !strings.Contains(outputJSON, `"name":"safe"`) {
-		t.Fatalf("structured connector output leaked sensitive field values: %s", outputJSON)
-	}
-}
-
-func TestConnectorActionApprovalRoutesDeclinePendingRequest(t *testing.T) {
-	fixture := newAPITestFixture(t)
-	store := connectortargets.NewStore(fixture.db)
-	token, err := fixture.tokens.Create(context.Background(), tokens.CreateRequest{Name: "codex"})
-	if err != nil {
-		t.Fatalf("create token: %v", err)
-	}
-	target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       token.ID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-	}); err != nil {
-		t.Fatalf("set connector permission: %v", err)
-	}
-	result, err := fixture.server.callConnectorAction(context.Background(), fixture.server.activeRuntime(), connectorActionCall{
-		Source:     commandRequestSourceMCP,
-		TokenID:    token.ID,
-		TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-		ActionName: postgresconnector.ActionQueryReadonly,
-		Input:      map[string]any{"sql": "select 1"},
-		Reason:     "smoke",
-	})
-	if err != nil {
-		t.Fatalf("call connector action: %v", err)
-	}
-
-	listResponse := performJSON(fixture.server.Handler(), http.MethodGet, "/api/connector-action-approvals?status=approval_pending", "", nil)
-	if listResponse.Code != http.StatusOK || !strings.Contains(listResponse.Body.String(), strconv.FormatInt(result.Request.ID, 10)) {
-		t.Fatalf("list connector approvals failed: %d %s", listResponse.Code, listResponse.Body.String())
-	}
-	detailResponse := performJSON(fixture.server.Handler(), http.MethodGet, "/api/connector-action-approvals/"+strconv.FormatInt(result.Request.ID, 10), "", nil)
-	if detailResponse.Code != http.StatusOK || !strings.Contains(detailResponse.Body.String(), "select 1") {
-		t.Fatalf("approval detail must expose exact pending preview: %d %s", detailResponse.Code, detailResponse.Body.String())
-	}
-	var encryptedPayload string
-	if err := fixture.db.QueryRow(`SELECT encrypted_payload_json FROM connector_action_requests WHERE id = ?`, result.Request.ID).Scan(&encryptedPayload); err != nil {
-		t.Fatalf("read encrypted approval payload: %v", err)
-	}
-	if _, err := fixture.db.Exec(`UPDATE connector_action_requests SET encrypted_payload_json = 'invalid' WHERE id = ?`, result.Request.ID); err != nil {
-		t.Fatalf("corrupt encrypted approval payload: %v", err)
-	}
-	redactedListResponse := performJSON(fixture.server.Handler(), http.MethodGet, "/api/connector-action-approvals?status=approval_pending", "", nil)
-	if redactedListResponse.Code != http.StatusOK {
-		t.Fatalf("approval list must not decrypt exact pending payloads: %d %s", redactedListResponse.Code, redactedListResponse.Body.String())
-	}
-	if _, err := fixture.db.Exec(`UPDATE connector_action_requests SET encrypted_payload_json = ? WHERE id = ?`, encryptedPayload, result.Request.ID); err != nil {
-		t.Fatalf("restore encrypted approval payload: %v", err)
-	}
-	declineResponse := performJSON(fixture.server.Handler(), http.MethodPost, "/api/connector-action-approvals/"+strconv.FormatInt(result.Request.ID, 10)+"/decline", "", declineConnectorActionApprovalRequest{UserNote: "not now"})
-	if declineResponse.Code != http.StatusOK || !strings.Contains(declineResponse.Body.String(), `"status":"declined"`) {
-		t.Fatalf("decline connector approval failed: %d %s", declineResponse.Code, declineResponse.Body.String())
-	}
-	mcpResponse := performJSON(fixture.server.Handler(), http.MethodGet, "/api/mcp/connector-action-requests/"+strconv.FormatInt(result.Request.ID, 10), token.TokenValue, nil)
-	if mcpResponse.Code != http.StatusOK || !strings.Contains(mcpResponse.Body.String(), `"status":"declined"`) || !strings.Contains(mcpResponse.Body.String(), "not now") {
-		t.Fatalf("mcp connector request should show decline: %d %s", mcpResponse.Code, mcpResponse.Body.String())
-	}
-}
-
-func TestConnectorActionApprovalRunUsesEncryptedInputNotRedactedDisplay(t *testing.T) {
-	fixture := newAPITestFixture(t)
-	store := connectortargets.NewStore(fixture.db)
-	token, err := fixture.tokens.Create(context.Background(), tokens.CreateRequest{Name: "codex"})
-	if err != nil {
-		t.Fatalf("create token: %v", err)
-	}
-	target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       token.ID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-	}); err != nil {
-		t.Fatalf("set connector permission: %v", err)
-	}
-	result, err := fixture.server.callConnectorAction(context.Background(), fixture.server.activeRuntime(), connectorActionCall{
-		Source:     commandRequestSourceMCP,
-		TokenID:    token.ID,
-		TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-		ActionName: postgresconnector.ActionQueryReadonly,
-		Input:      map[string]any{"sql": "select 'password=super-secret' as value"},
-		Reason:     "smoke",
-	})
-	if err != nil {
-		t.Fatalf("call connector action: %v", err)
-	}
-	if !strings.Contains(fmt.Sprint(result.Request.Input["sql"]), "password=[REDACTED]") {
-		t.Fatalf("display input should be redacted: %#v", result.Request.Input)
-	}
-
-	runResponse := performJSON(fixture.server.Handler(), http.MethodPost, "/api/connector-action-approvals/"+strconv.FormatInt(result.Request.ID, 10)+"/run", "", runConnectorActionApprovalRequest{})
-	if runResponse.Code != http.StatusOK {
-		t.Fatalf("approval run should not fail stale because display input was redacted: %d %s", runResponse.Code, runResponse.Body.String())
-	}
-	finished, err := store.GetActionRequest(context.Background(), result.Request.ID)
-	if err != nil {
-		t.Fatalf("get finished connector request: %v", err)
-	}
-	if finished.Status == connectors.ResultStale {
-		t.Fatalf("request should not become stale from redacted display input: %#v", finished)
-	}
-}
-
-func TestConnectorActionApprovalRunDeliversUserNote(t *testing.T) {
-	fixture := newAPITestFixture(t)
-	store := connectortargets.NewStore(fixture.db)
-	token, err := fixture.tokens.Create(context.Background(), tokens.CreateRequest{Name: "codex"})
-	if err != nil {
-		t.Fatalf("create token: %v", err)
-	}
-	target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       token.ID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-	}); err != nil {
-		t.Fatalf("set connector permission: %v", err)
-	}
-	result, err := fixture.server.callConnectorAction(context.Background(), fixture.server.activeRuntime(), connectorActionCall{
-		Source:     commandRequestSourceMCP,
-		TokenID:    token.ID,
-		TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-		ActionName: postgresconnector.ActionQueryReadonly,
-		Input:      map[string]any{"sql": "select 1"},
-		Reason:     "smoke",
-	})
-	if err != nil {
-		t.Fatalf("call connector action: %v", err)
-	}
-
-	runResponse := performJSON(fixture.server.Handler(), http.MethodPost, "/api/connector-action-approvals/"+strconv.FormatInt(result.Request.ID, 10)+"/run", "", runConnectorActionApprovalRequest{UserNote: "only inspect metadata"})
-	if runResponse.Code != http.StatusOK {
-		t.Fatalf("approval run failed: %d %s", runResponse.Code, runResponse.Body.String())
-	}
-	var queued string
-	if err := fixture.db.QueryRow(`
-		SELECT message
-		FROM message_queue
-		WHERE token_id = ? AND direction = 'user_to_ai'
-		ORDER BY id DESC
-		LIMIT 1`,
-		token.ID,
-	).Scan(&queued); err != nil {
-		t.Fatalf("read queued approval note: %v", err)
-	}
-	if !strings.Contains(queued, "only inspect metadata") {
-		t.Fatalf("queued note = %q", queued)
-	}
-}
-
-func TestConnectorActionApprovalRunMarksDriftStale(t *testing.T) {
-	fixture := newAPITestFixture(t)
-	store := connectortargets.NewStore(fixture.db)
-	token, err := fixture.tokens.Create(context.Background(), tokens.CreateRequest{Name: "codex"})
-	if err != nil {
-		t.Fatalf("create token: %v", err)
-	}
-	target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       token.ID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-	}); err != nil {
-		t.Fatalf("set connector permission: %v", err)
-	}
-	result, err := fixture.server.callConnectorAction(context.Background(), fixture.server.activeRuntime(), connectorActionCall{
-		Source:     commandRequestSourceMCP,
-		TokenID:    token.ID,
-		TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-		ActionName: postgresconnector.ActionQueryReadonly,
-		Input:      map[string]any{"sql": "select 1"},
-		Reason:     "smoke",
-	})
-	if err != nil {
-		t.Fatalf("call connector action: %v", err)
-	}
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       token.ID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    postgresconnector.ActionQueryReadonly,
-		ExecutionRule: connectortargets.ActionPermissionBlocked,
-	}); err != nil {
-		t.Fatalf("block connector permission: %v", err)
-	}
-
-	runResponse := performJSON(fixture.server.Handler(), http.MethodPost, "/api/connector-action-approvals/"+strconv.FormatInt(result.Request.ID, 10)+"/run", "", runConnectorActionApprovalRequest{})
-	if runResponse.Code != http.StatusConflict || !strings.Contains(runResponse.Body.String(), "fresh request") {
-		t.Fatalf("expected stale conflict, got %d %s", runResponse.Code, runResponse.Body.String())
-	}
-	stale, err := store.GetActionRequest(context.Background(), result.Request.ID)
-	if err != nil {
-		t.Fatalf("get stale connector request: %v", err)
-	}
-	if stale.Status != connectors.ResultStale {
-		t.Fatalf("status = %q", stale.Status)
-	}
-	if stale.ApprovalContextDrift != "permission" {
-		t.Fatalf("approval drift = %q", stale.ApprovalContextDrift)
-	}
-}
-
-func TestConnectorActionApprovalRunMarksPrepareFailureStale(t *testing.T) {
-	fixture := newAPITestFixture(t)
-	store := connectortargets.NewStore(fixture.db)
-	token, err := fixture.tokens.Create(context.Background(), tokens.CreateRequest{Name: "codex"})
-	if err != nil {
-		t.Fatalf("create token: %v", err)
-	}
-	target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-	badAction := "missing_action"
-	if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-		TokenID:       token.ID,
-		TargetID:      target.ID,
-		ProfileID:     profile.ID,
-		ActionName:    badAction,
-		ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-	}); err != nil {
-		t.Fatalf("set connector permission: %v", err)
-	}
-	encryptedPayload, err := fixture.server.activeRuntime().vault.EncryptJSON(connectorActionExecutionEnvelope{
-		Input:   map[string]any{},
-		Payload: map[string]any{},
-	})
-	if err != nil {
-		t.Fatalf("encrypt pending payload: %v", err)
-	}
-	request, err := store.InsertActionRequest(context.Background(), connectortargets.InsertActionRequestInput{
-		TokenID:              &token.ID,
-		TargetID:             target.ID,
-		ProfileID:            profile.ID,
-		ConnectorKind:        postgresconnector.Kind,
-		ActionName:           badAction,
-		Source:               commandRequestSourceMCP,
-		Input:                map[string]any{},
-		EncryptedPayloadJSON: encryptedPayload,
-		Status:               connectors.ResultApprovalPending,
-		ApprovalContextHash:  "old-context",
-	})
-	if err != nil {
-		t.Fatalf("insert pending connector request: %v", err)
-	}
-	if err := historypkg.NewStore(fixture.db).SyncConnectorActionRequest(context.Background(), request.ID); err != nil {
-		t.Fatalf("sync pending connector request: %v", err)
-	}
-
-	runResponse := performJSON(fixture.server.Handler(), http.MethodPost, "/api/connector-action-approvals/"+strconv.FormatInt(request.ID, 10)+"/run", "", runConnectorActionApprovalRequest{})
-	if runResponse.Code != http.StatusConflict || !strings.Contains(runResponse.Body.String(), "fresh request") {
-		t.Fatalf("expected prepare drift conflict, got %d %s", runResponse.Code, runResponse.Body.String())
-	}
-	stale, err := store.GetActionRequest(context.Background(), request.ID)
-	if err != nil {
-		t.Fatalf("get stale connector request: %v", err)
-	}
-	if stale.Status != connectors.ResultStale || !strings.Contains(stale.Error, "fresh request") {
-		t.Fatalf("request should be stale with fresh-request error: %#v", stale)
-	}
-	if stale.ApprovalContextDrift != "target_or_action" {
-		t.Fatalf("approval drift = %q", stale.ApprovalContextDrift)
-	}
-	var historyStatus string
-	if err := fixture.db.QueryRow(`
-		SELECT status
-		FROM history_entries
-		WHERE source_ref_type = 'connector_action_request' AND source_ref_id = ?`,
-		request.ID,
-	).Scan(&historyStatus); err != nil {
-		t.Fatalf("read synced history: %v", err)
-	}
-	if historyStatus != string(connectors.ResultStale) {
-		t.Fatalf("history status = %q", historyStatus)
-	}
-}
-
-func TestConnectorTargetAndProfileUpdatesStalePendingApprovals(t *testing.T) {
-	fixture := newAPITestFixture(t)
-	ctx := context.Background()
-	store := connectortargets.NewStore(fixture.db)
-	token, err := fixture.tokens.Create(ctx, tokens.CreateRequest{Name: "codex"})
-	if err != nil {
-		t.Fatalf("create token: %v", err)
-	}
-	target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-	pendingTarget, err := store.InsertActionRequest(ctx, connectortargets.InsertActionRequestInput{
-		TokenID:              &token.ID,
-		TargetID:             target.ID,
-		ProfileID:            profile.ID,
-		ConnectorKind:        postgresconnector.Kind,
-		ActionName:           postgresconnector.ActionQueryReadonly,
-		Input:                map[string]any{"sql": "select 1"},
-		EncryptedPayloadJSON: "encrypted-payload",
-		Status:               connectors.ResultApprovalPending,
-	})
-	if err != nil {
-		t.Fatalf("insert pending target request: %v", err)
-	}
-	running, err := store.InsertActionRequest(ctx, connectortargets.InsertActionRequestInput{
-		TokenID:              &token.ID,
-		TargetID:             target.ID,
-		ProfileID:            profile.ID,
-		ConnectorKind:        postgresconnector.Kind,
-		ActionName:           postgresconnector.ActionQueryReadonly,
-		Input:                map[string]any{"sql": "select pg_sleep(10)"},
-		EncryptedPayloadJSON: "encrypted-payload",
-		Status:               connectors.ResultRunning,
-	})
-	if err != nil {
-		t.Fatalf("insert running request: %v", err)
-	}
-	updateTarget := performJSON(fixture.server.Handler(), http.MethodPut, "/api/connector-targets/"+strconv.FormatInt(target.ID, 10), "", updateConnectorTargetRequest{
-		Name: "main-db-renamed",
-		Config: map[string]any{
-			"connection_mode": "direct",
-			"host":            "127.0.0.1",
-			"port":            5432,
-			"database":        "app",
-			"ssl_mode":        "prefer",
-		},
-	})
-	if updateTarget.Code != http.StatusOK {
-		t.Fatalf("update connector target failed: %d %s", updateTarget.Code, updateTarget.Body.String())
-	}
-	gotPendingTarget, err := store.GetActionRequest(ctx, pendingTarget.ID)
-	if err != nil {
-		t.Fatalf("get pending target request: %v", err)
-	}
-	if gotPendingTarget.Status != connectors.ResultStale || gotPendingTarget.ApprovalContextDrift != "target" {
-		t.Fatalf("pending target request should be stale after target update: %#v", gotPendingTarget)
-	}
-	gotRunning, err := store.GetActionRequest(ctx, running.ID)
-	if err != nil {
-		t.Fatalf("get running request: %v", err)
-	}
-	if gotRunning.Status != connectors.ResultRunning {
-		t.Fatalf("running request should not be stale from target update: %#v", gotRunning)
-	}
-
-	pendingProfile, err := store.InsertActionRequest(ctx, connectortargets.InsertActionRequestInput{
-		TokenID:              &token.ID,
-		TargetID:             target.ID,
-		ProfileID:            profile.ID,
-		ConnectorKind:        postgresconnector.Kind,
-		ActionName:           postgresconnector.ActionGetSchemas,
-		Input:                map[string]any{},
-		EncryptedPayloadJSON: "encrypted-payload",
-		Status:               connectors.ResultApprovalPending,
-	})
-	if err != nil {
-		t.Fatalf("insert pending profile request: %v", err)
-	}
-	updateProfile := performJSON(fixture.server.Handler(), http.MethodPut, "/api/connector-targets/"+strconv.FormatInt(target.ID, 10)+"/profiles/"+strconv.FormatInt(profile.ID, 10), "", updateConnectorCredentialProfileRequest{
-		Kind:  "username_password",
-		Label: "readonly-renamed",
-		Public: map[string]any{
-			"username": "app_readonly",
-		},
-		RiskLabel: "read-only",
-	})
-	if updateProfile.Code != http.StatusOK {
-		t.Fatalf("update connector profile failed: %d %s", updateProfile.Code, updateProfile.Body.String())
-	}
-	gotPendingProfile, err := store.GetActionRequest(ctx, pendingProfile.ID)
-	if err != nil {
-		t.Fatalf("get pending profile request: %v", err)
-	}
-	if gotPendingProfile.Status != connectors.ResultStale || gotPendingProfile.ApprovalContextDrift != "profile" {
-		t.Fatalf("pending profile request should be stale after profile update: %#v", gotPendingProfile)
-	}
-}
-
-func TestConnectorActionApprovalRunRequiresCurrentToken(t *testing.T) {
-	for name, mutate := range map[string]func(*testing.T, apiTestFixture, tokens.CreateResponse){
-		"revoked": func(t *testing.T, fixture apiTestFixture, token tokens.CreateResponse) {
-			t.Helper()
-			if _, err := fixture.tokens.Revoke(context.Background(), token.ID); err != nil {
-				t.Fatalf("revoke token: %v", err)
-			}
-		},
-		"expired": func(t *testing.T, fixture apiTestFixture, token tokens.CreateResponse) {
-			t.Helper()
-			if _, err := fixture.db.Exec(`UPDATE api_tokens SET expires_at = ? WHERE id = ?`, time.Now().UTC().Add(-time.Minute).Format(time.RFC3339), token.ID); err != nil {
-				t.Fatalf("expire token: %v", err)
-			}
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			fixture := newAPITestFixture(t)
-			store := connectortargets.NewStore(fixture.db)
-			token, err := fixture.tokens.Create(context.Background(), tokens.CreateRequest{Name: "codex"})
-			if err != nil {
-				t.Fatalf("create token: %v", err)
-			}
-			target, profile := createAPITestPostgresTargetProfile(t, store, fixture.server.activeRuntime().vault)
-			if err := store.SetActionPermission(context.Background(), connectortargets.SetActionPermissionInput{
-				TokenID:       token.ID,
-				TargetID:      target.ID,
-				ProfileID:     profile.ID,
-				ActionName:    postgresconnector.ActionQueryReadonly,
-				ExecutionRule: connectortargets.ActionPermissionApprovalRequired,
-			}); err != nil {
-				t.Fatalf("set connector permission: %v", err)
-			}
-			result, err := fixture.server.callConnectorAction(context.Background(), fixture.server.activeRuntime(), connectorActionCall{
-				Source:     commandRequestSourceMCP,
-				TokenID:    token.ID,
-				TargetRef:  connectortargets.ConnectorTargetRef(postgresconnector.Kind, target.ID, profile.ID),
-				ActionName: postgresconnector.ActionQueryReadonly,
-				Input:      map[string]any{"sql": "select 1"},
-				Reason:     "smoke",
-			})
-			if err != nil {
-				t.Fatalf("call connector action: %v", err)
-			}
-
-			mutate(t, fixture, token)
-			runResponse := performJSON(fixture.server.Handler(), http.MethodPost, "/api/connector-action-approvals/"+strconv.FormatInt(result.Request.ID, 10)+"/run", "", runConnectorActionApprovalRequest{})
-			if runResponse.Code != http.StatusConflict || !strings.Contains(runResponse.Body.String(), "fresh request") {
-				t.Fatalf("expected stale conflict, got %d %s", runResponse.Code, runResponse.Body.String())
-			}
-			stale, err := store.GetActionRequest(context.Background(), result.Request.ID)
-			if err != nil {
-				t.Fatalf("get stale connector request: %v", err)
-			}
-			if stale.Status != connectors.ResultStale {
-				t.Fatalf("status = %q", stale.Status)
-			}
-		})
-	}
-}
-
-func openAPITestDB(t *testing.T) *sql.DB {
-	t.Helper()
-	database, err := dbpkg.OpenEncrypted(filepath.Join(t.TempDir(), "test.db"), "test-password")
-	if err != nil {
-		t.Fatalf("open db: %v", err)
-	}
-	t.Cleanup(func() {
-		_ = database.Close()
-	})
-	return database
-}
-
-func connectorActionTestRuntime(t *testing.T, database *sql.DB, secretVault *vault.Vault) *databaseRuntime {
-	t.Helper()
-	return &databaseRuntime{
-		database: database,
-		vault:    secretVault,
-		tokens:   tokens.NewStore(database),
-		registry: testConnectorRegistry(t),
-	}
-}
-
-func openAPITestVault(t *testing.T) *vault.Vault {
-	t.Helper()
-	secretVault, err := vault.New("test-gateway-secret")
-	if err != nil {
-		t.Fatalf("create vault: %v", err)
-	}
-	return secretVault
-}
-
-func createAPITestPostgresTargetProfile(t *testing.T, store *connectortargets.Store, secretVault *vault.Vault) (connectortargets.Target, connectortargets.CredentialProfile) {
-	t.Helper()
-	ctx := context.Background()
-	target, err := store.CreateTarget(ctx, connectortargets.CreateTargetInput{
-		ConnectorKind: postgresconnector.Kind,
-		Name:          "main-db",
-		Config: map[string]any{
-			"connection_mode": "direct",
-			"host":            "127.0.0.1",
-			"port":            5432,
-			"database":        "app",
-			"ssl_mode":        "disable",
-		},
-	})
-	if err != nil {
-		t.Fatalf("create postgres target: %v", err)
-	}
-	encryptedSecret, err := secretVault.EncryptJSON(map[string]any{"password": "secret"})
-	if err != nil {
-		t.Fatalf("encrypt profile secret: %v", err)
-	}
-	profile, err := store.CreateCredentialProfile(ctx, connectortargets.CreateCredentialProfileInput{
-		TargetID:            target.ID,
-		ConnectorKind:       postgresconnector.Kind,
-		Kind:                "username_password",
-		Label:               "readonly",
-		Public:              map[string]any{"username": "app_readonly"},
-		EncryptedSecretJSON: encryptedSecret,
-	})
-	if err != nil {
-		t.Fatalf("create postgres profile: %v", err)
-	}
-	return target, profile
-}
-
-func insertAPITestToken(t *testing.T, database *sql.DB) int64 {
-	t.Helper()
-	now := time.Now().UTC().Format(time.RFC3339)
-	result, err := database.Exec(`
-		INSERT INTO api_tokens (name, token_hash, token_prefix, created_at, updated_at)
-		VALUES ('connector-codex', 'connector-hash', 'aip_conn', ?, ?)`,
-		now,
-		now,
-	)
-	if err != nil {
-		t.Fatalf("insert token: %v", err)
-	}
-	id, err := result.LastInsertId()
-	if err != nil {
-		t.Fatalf("token id: %v", err)
-	}
-	if _, err := database.Exec(`
-		INSERT INTO token_project_scopes (token_id, project_id, enabled, created_at, updated_at)
-		SELECT ?, id, 1, ?, ? FROM projects WHERE status = 'active'`, id, now, now); err != nil {
-		t.Fatalf("insert token project scopes: %v", err)
-	}
-	return id
-}
-
-const localActionTestConnectorKind = "localtest"
-
-type localActionTestConnector struct{}
-
-func (localActionTestConnector) ValidateTargetConfig(config map[string]any) error {
-	if value, _ := config["semantic_error"].(bool); value {
-		return fmt.Errorf("semantic validation fixture")
-	}
-	return nil
-}
-
-func (localActionTestConnector) Kind() string {
-	return localActionTestConnectorKind
-}
-
-func (localActionTestConnector) Label() string {
-	return "Local Test"
-}
-
-func (localActionTestConnector) Version() string {
-	return "0.1"
-}
-
-func (localActionTestConnector) TargetSchema() connectors.Schema {
-	return connectors.Schema{}
-}
-
-func (localActionTestConnector) CredentialSchemas() []connectors.CredentialSchema {
-	return []connectors.CredentialSchema{{Kind: "default", Label: "Default", Schema: connectors.Schema{}}}
-}
-
-func (localActionTestConnector) GetHelp(context.Context, connectors.TargetView) (connectors.ConnectorHelp, error) {
-	return connectors.ConnectorHelp{
-		Title:       "Local test target",
-		Summary:     "Test connector.",
-		Connector:   "Local Test",
-		ConnectorID: localActionTestConnectorKind,
-	}, nil
-}
-
-func (localActionTestConnector) GetActionList(context.Context, connectors.TargetView, connectors.CredentialProfileView) ([]connectors.ActionDefinition, error) {
-	return []connectors.ActionDefinition{{
-		Name:        "echo",
-		Label:       "Echo",
-		Description: "Echo one value.",
-		Risk:        connectors.RiskRead,
-		InputSchema: connectors.Schema{Fields: []connectors.Field{{
-			Name:     "value",
-			Label:    "Value",
-			Type:     connectors.FieldString,
-			Required: true,
-		}}},
-	}}, nil
-}
-
-func (localActionTestConnector) PrepareAction(_ context.Context, req connectors.ActionRequest) (connectors.PreparedAction, error) {
-	value, _ := req.Input["value"].(string)
-	return connectors.PreparedAction{
-		ConnectorKind: localActionTestConnectorKind,
-		TargetRef:     req.Target.Ref,
-		ProfileID:     req.Profile.ID,
-		ActionName:    req.ActionName,
-		Risk:          connectors.RiskRead,
-		Title:         "Echo value",
-		Summary:       "Echo one test value",
-		Preview:       map[string]any{"value": value},
-		Payload:       map[string]any{"value": value},
-		ContextMaterial: map[string]any{
-			"value": value,
-		},
-	}, nil
-}
-
-func (localActionTestConnector) ExecuteAction(_ context.Context, _ connectors.RuntimeContext, action connectors.PreparedAction) (connectors.ActionResult, error) {
-	if action.Payload["value"] == "classified-error" {
-		return connectors.ActionResult{}, connectors.ClassifyError("fixture_failure", fmt.Errorf("fixture failed"))
-	}
-	result := connectors.ActionResult{
-		Status:      connectors.ResultCompleted,
-		Output:      map[string]any{"echo": action.Payload["value"]},
-		DisplayText: fmt.Sprint(action.Payload["value"]),
-	}
-	if action.Payload["value"] == "with-handle" {
-		result.Handles = connectors.ActionHandles{SessionID: 123, SessionGeneration: 456}
-	}
-	if action.Payload["value"] == "incomplete-handle" {
-		result.Handles = connectors.ActionHandles{SessionID: 123}
-	}
-	return result, nil
 }
